separate out onnx ctx model related code to onnx_ctx_model_helper file
