+++ onnxruntime/core/providers/qnn/builder/qnn_model_wrapper.h
@@ -241,64 +241,5 @@ class QnnModelWrapper {
   QnnBackendType qnn_backend_type_ = QnnBackendType::CPU;
 };  // QnnModelWrapper
 
-class QnnCacheModelHandler {
- public:
-  QnnCacheModelHandler() {
-  }
-  ORT_DISALLOW_COPY_ASSIGNMENT_AND_MOVE(QnnCacheModelHandler);
-
-  bool IsQnnCtxModel(const onnxruntime::GraphViewer& graph_viewer) {
-    if (is_qnn_ctx_model_) {
-      return is_qnn_ctx_model_;
-    }
-    // It's a Onnx model with Qnn context cache binary if it only has a node with EPCache type
-    if (graph_viewer.NumberOfNodes() == 1 && graph_viewer.Nodes().begin()->OpType() == "EPCache") {
-      is_qnn_ctx_model_ = true;
-    }
-    return is_qnn_ctx_model_;
-  }
-
-  bool GetIsQnnCtxModel() const {
-    return is_qnn_ctx_model_;
-  }
-
-  Status GetEpEngineCache(const onnxruntime::GraphViewer& graph_viewer,
-                          const std::string& onnx_model_path,
-                          bool is_ctx_cache_file_exist,
-                          std::string& ep_engine_cache,
-                          const logging::Logger& logger) const {
-    if (is_qnn_ctx_model_) {
-      ORT_RETURN_IF_ERROR(GetEpEngineCacheFromGraph(graph_viewer, ep_engine_cache));
-    } else if (is_ctx_cache_file_exist) {
-      ORT_RETURN_IF_ERROR(GetEpEngineCacheFromModel(onnx_model_path, ep_engine_cache, logger));
-    }
-
-    return Status::OK();
-  }
-
- private:
-  Status GetEpEngineCacheFromModel(const std::string& onnx_model_path,
-                                   std::string& ep_engine_cache,
-                                   const logging::Logger& logger) const {
-    using namespace onnxruntime;
-    std::shared_ptr<Model> model;
-    ORT_RETURN_IF_ERROR(Model::Load(ToPathString(onnx_model_path), model, {}, logger));
-    const auto& graph = model->MainGraph();
-    ORT_RETURN_IF_ERROR(GetEpEngineCacheFromGraph(GraphViewer(graph), ep_engine_cache));
-
-    return Status::OK();
-  }
-
-  Status GetEpEngineCacheFromGraph(const onnxruntime::GraphViewer& graph_viewer, std::string& ep_engine_cache) const {
-    const auto& node = graph_viewer.Nodes().begin();
-    NodeAttrHelper node_helper(*node);
-    ep_engine_cache = node_helper.Get("ep_engine_cache", ep_engine_cache);
-    return Status::OK();
-  }
-
- private:
-  bool is_qnn_ctx_model_ = false;
-}; // QnnCacheModelHandler
-
 }  // namespace qnn
 }  // namespace onnxruntime
